requirements.txt
django==3.1.6
psycopg2==2.8.6
Pillow==8.1.0
djangorestframework==3.12.2